requirements.txt
streamlit
requests
Pillow
pytesseract
opencv-python
pandas
numpy
python-dotenv
